chore: refactor creating the folder mori and moving files, server.py and llm_manager.py into it
